python/ai-agent-builders/requirements.txt
# Common
python-dotenv==1.0.1

# LangChain
langchain==0.3.25
langchain-core==0.3.59
langchain-google-genai==2.1.4

# AutoGen
autogen-agentchat==0.4.9.3
autogen-core==0.4.9.3
autogen-ext==0.4.9.3

# CrewAI
crewai==0.119.0
